+++ jiber-client/src/index.ts
@@ -1,1 +1,1 @@
-export * from './store/jiber-client'
+export * from './jiber-client'
