+++ jiber-client/src/interfaces/index.ts
@@ -1,4 +1,3 @@
 export * from './action-creators'
 export * from './action'
-export * from './client-settings'
 export * from './reducer'
