+++ jiber-client/src/packet.test.ts
