+++ jiber-client/src/packet.ts
